Adds alternative method to generate display name and email
Allows user to ignore cas attributes when DisplayName and/or Email are not available in the CAS ticket response.  By specifying $wgCas_ExcludeAttributes=true the DisplayName is set to the username and the email is generated from the username and $wgCas_EmailDomain.  In other words, the email is <username>@<$wgCas_EmailDomain>